Rename file (#999)
