+++ packages/patrol/lib/patrol.dart
@@ -8,5 +8,5 @@ export 'src/binding.dart';
 // ignore: invalid_export_of_internal_element
 export 'src/common.dart';
 // ignore: invalid_export_of_internal_element
-export 'src/custom_finders/patrol_integration_test.dart';
+export 'src/custom_finders/patrol_integration_tester.dart';
 export 'src/native/native.dart';
